requirements.txt
streamlit
torch
transformers
diffusers
accelerate
sentencepiece
huggingface-hub
pillow
numpy
scipy
requests
tqdm
